# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: games
    description: Educational games
  - name: events
    description: Project events and outreach
  - name: results
    description: Research results and feedback
  - name: documents
    description: Academic documents (PDFs, reports, articles)
  - name: team
    description: Team members
  - name: stats
    description: Dashboard summaries
  - name: storage
    description: Object storage uploads
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
  /auth/me:
    get:
      operationId: getAuthStatus
      tags: [auth]
      summary: Get current auth status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"

  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List all research results
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      summary: Create a new research result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResultBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /results/{id}:
    delete:
      operationId: deleteResult
      tags: [results]
      summary: Delete a result
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List academic documents
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
            enum: [report, article, certificate, edital, other]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /team:
    get:
      operationId: listTeam
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Add a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Remove a team member
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: Dashboard overview stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"
  /stats/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Recent project activity feed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    LoginBody:
      type: object
      properties:
        password:
          type: string
      required: [password]
    AuthStatus:
      type: object
      properties:
        isAdmin:
          type: boolean
      required: [isAdmin]

    Game:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        kind:
          type: string
          enum: [virtual, board, hybrid]
        playUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, kind, tags, createdAt]
    CreateGameBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        kind:
          type: string
          enum: [virtual, board, hybrid]
        playUrl:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        tags:
          type: array
          items: { type: string }
      required: [title, description, kind]
    UpdateGameBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        kind:
          type: string
          enum: [virtual, board, hybrid]
        playUrl: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }

    Event:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        location: { type: string }
        eventDate: { type: string, format: date }
        kind:
          type: string
          enum: [oficina, apresentacao, acao_social, evento_academico]
        imageUrl: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, location, eventDate, kind, createdAt]
    CreateEventBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        location: { type: string }
        eventDate: { type: string, format: date }
        kind:
          type: string
          enum: [oficina, apresentacao, acao_social, evento_academico]
        imageUrl: { type: string, nullable: true }
      required: [title, description, location, eventDate, kind]
    UpdateEventBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        location: { type: string }
        eventDate: { type: string, format: date }
        kind:
          type: string
          enum: [oficina, apresentacao, acao_social, evento_academico]
        imageUrl: { type: string, nullable: true }

    Result:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        summary: { type: string }
        participants: { type: integer }
        satisfactionScore:
          type: number
          format: float
          nullable: true
        learningScore:
          type: number
          format: float
          nullable: true
        methodology: { type: string }
        gameId:
          type: integer
          nullable: true
        createdAt: { type: string, format: date-time }
      required: [id, title, summary, participants, methodology, createdAt]
    CreateResultBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        summary: { type: string }
        participants: { type: integer, minimum: 0 }
        satisfactionScore:
          type: number
          format: float
          nullable: true
        learningScore:
          type: number
          format: float
          nullable: true
        methodology: { type: string }
        gameId:
          type: integer
          nullable: true
      required: [title, summary, participants, methodology]

    Document:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category:
          type: string
          enum: [report, article, certificate, edital, other]
        program:
          type: string
          enum: [pibex, pic, geral]
        fileUrl: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, category, program, fileUrl, createdAt]
    CreateDocumentBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        category:
          type: string
          enum: [report, article, certificate, edital, other]
        program:
          type: string
          enum: [pibex, pic, geral]
        fileUrl: { type: string, minLength: 1 }
      required: [title, description, category, program, fileUrl]

    TeamMember:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        course: { type: string }
        program:
          type: string
          enum: [pibex, pic, ambos, orientador]
        bio:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        sortOrder: { type: integer }
      required: [id, name, role, course, program, sortOrder]
    CreateTeamMemberBody:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        role: { type: string }
        course: { type: string }
        program:
          type: string
          enum: [pibex, pic, ambos, orientador]
        bio: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true }
        sortOrder: { type: integer }
      required: [name, role, course, program]
    UpdateTeamMemberBody:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        course: { type: string }
        program:
          type: string
          enum: [pibex, pic, ambos, orientador]
        bio: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true }
        sortOrder: { type: integer }

    StatsOverview:
      type: object
      properties:
        totalGames: { type: integer }
        totalEvents: { type: integer }
        totalParticipants: { type: integer }
        totalDocuments: { type: integer }
        totalTeamMembers: { type: integer }
        avgSatisfaction:
          type: number
          format: float
          nullable: true
      required:
        - totalGames
        - totalEvents
        - totalParticipants
        - totalDocuments
        - totalTeamMembers

    ActivityItem:
      type: object
      properties:
        id: { type: string }
        kind:
          type: string
          enum: [game, event, result, document]
        title: { type: string }
        subtitle: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, kind, title, subtitle, createdAt]

    RequestUploadUrlBody:
      type: object
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
      required: [name, size, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [uploadURL, objectPath]