# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentCategory } from "./documentCategory";
import type { DocumentProgram } from "./documentProgram";

export interface Document {
  id: number;
  title: string;
  description: string;
  category: DocumentCategory;
  program: DocumentProgram;
  fileUrl: string;
  createdAt: Date;
}