# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameKind } from "./gameKind";

export interface Game {
  id: number;
  title: string;
  description: string;
  kind: GameKind;
  playUrl?: string | null;
  imageUrl?: string | null;
  tags: string[];
  createdAt: Date;
}